+++ src/gurubase-frontend/src/components/NewGuru.jsx
@@ -4,19 +4,12 @@ import { useUser } from "@auth0/nextjs-auth0/client";
 import { zodResolver } from "@hookform/resolvers/zod";
 import {
   AlertTriangle,
-  Clock,
-  Edit,
   Info,
   LinkIcon,
   LoaderCircle,
   Lock,
-  MoreVertical,
-  RotateCw,
-  Unlock,
-  Upload,
-  Check
+  Unlock
 } from "lucide-react";
-import Image from "next/image";
 import { redirect } from "next/navigation";
 import { useCallback, useEffect, useRef, useState } from "react";
 import * as React from "react";
@@ -39,59 +32,12 @@ import {
   getIntegrationDetails // <-- Import getIntegrationDetails
 } from "@/app/actions";
 import { CustomToast } from "@/components/CustomToast";
-import {
-  LogosYoutubeIcon,
-  LucideSquareArrowOutUpRight,
-  SolarFileTextBold,
-  SolarGalleryAddBold,
-  SolarInfoCircleBold,
-  SolarTrashBinTrashBold,
-  SolarVideoLibraryBold,
-  JiraIcon
-} from "@/components/Icons";
+import { LucideSquareArrowOutUpRight } from "@/components/Icons";
 // Import the new component
 import SourceDialog from "@/components/NewEditGuru/SourceDialog";
-import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger
-} from "@/components/ui/dropdown-menu";
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle
-} from "@/components/ui/modal-dialog";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue
-} from "@/components/ui/select";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow
-} from "@/components/ui/table";
+import { Form } from "@/components/ui/form";
 import { useMediaQuery } from "@/hooks/useMediaQuery";
 import { cn } from "@/lib/utils";
 import {
@@ -106,14 +52,14 @@ import {
   TooltipProvider,
   TooltipTrigger
 } from "./ui/tooltip";
-import { HeaderTooltip } from "@/components/ui/header-tooltip";
 import { useCrawler } from "@/hooks/useCrawler";
 import { DeleteConfirmationModal } from "@/components/NewEditGuru/DeleteConfirmationModal";
 import { LongUpdatesIndicator } from "@/components/NewEditGuru/LongUpdatesIndicator";
 import { JiraIntegrationModal } from "@/components/NewEditGuru/JiraIntegrationModal";
 import { PendingChangesIndicator } from "@/components/NewEditGuru/PendingChangesIndicator";
 import { GithubSourceSection } from "@/components/NewEditGuru/GithubSourceSection";
-import { SourceActions } from "@/components/NewEditGuru/SourceActions";
+import { GuruDetailsSection } from "@/components/NewEditGuru/GuruDetailsSection"; // Import new component
+import { SourcesTableSection } from "@/components/NewEditGuru/SourcesTableSection"; // Import new component
 
 const formSchema = z.object({
   guruName: z
@@ -1747,8 +1693,6 @@ export default function NewGuru({ guruData, isProcessing }) {
     return () => window.removeEventListener("beforeunload", handleBeforeUnload);
   }, [hasFormChanged]);
 
-  // First, let's modify the renderBadges function to handle badge clicks
-
   const handlePrivacyBadgeClick = (e, source) => {
     e.preventDefault();
     e.stopPropagation();
@@ -1815,183 +1759,6 @@ export default function NewGuru({ guruData, isProcessing }) {
         ]
       };
     });
-  };
-
-  // Update the renderBadges function to include the click handler
-  const renderBadges = (source) => {
-    // For PDF files (existing code)
-    if (source?.type?.toLowerCase() === "pdf") {
-      return (
-        <div className="flex items-center gap-1">
-          {(() => {
-            let badgeProps = {
-              className: cn(
-                "flex items-center rounded-full gap-1 px-2 py-1 text-body4 font-medium cursor-pointer",
-                isSourcesProcessing && "pointer-events-none opacity-50"
-              ),
-              variant: "secondary"
-            };
-
-            switch (source.private) {
-              case true:
-                badgeProps.icon = Lock;
-                badgeProps.iconColor = "text-gray-500";
-                badgeProps.text = "Private";
-                break;
-              default:
-                badgeProps.icon = Unlock;
-                badgeProps.iconColor = "text-blue-base";
-                badgeProps.text = "Public";
-                break;
-            }
-
-            return (
-              <div className="flex items-center gap-1">
-                <div
-                  key={source.id}
-                  role="button"
-                  tabIndex={0}
-                  onClick={(e) => handlePrivacyBadgeClick(e, source)}
-                  onKeyDown={(e) => {
-                    if (e.key === "Enter" || e.key === " ") {
-                      handlePrivacyBadgeClick(e, source);
-                    }
-                  }}>
-                  <Badge {...badgeProps}>
-                    <badgeProps.icon
-                      className={cn("h-3 w-3", badgeProps.iconColor)}
-                    />
-                    {badgeProps.text}
-                  </Badge>
-                </div>
-                <div className="relative flex items-center">
-                  <TooltipProvider>
-                    <Tooltip delayDuration={0}>
-                      <TooltipTrigger className="cursor-pointer hover:text-gray-600 transition-colors flex items-center">
-                        <Info className="h-3.5 w-3.5 text-gray-400" />
-                      </TooltipTrigger>
-                      <TooltipContent
-                        align="center"
-                        className="rounded-lg shadow-lg border p-3 md:bg-[#1B242D] md:text-white bg-background"
-                        side="top"
-                        sideOffset={8}>
-                        <div
-                          className="absolute w-4 h-4 border-l border-t md:bg-[#1B242D] bg-background"
-                          style={{
-                            bottom: "-8px",
-                            left: "50%",
-                            transform: "translateX(-50%) rotate(225deg)",
-                            borderColor: "inherit"
-                          }}
-                        />
-                        <p className="text-center relative font-inter px-2 text-xs font-medium">
-                          {source.private
-                            ? "This resource will be listed but not linked as a question reference."
-                            : "This resource will be listed and linked as a question reference."}
-                        </p>
-                      </TooltipContent>
-                    </Tooltip>
-                  </TooltipProvider>
-                </div>
-              </div>
-            );
-          })()}
-        </div>
-      );
-    }
-
-    // Group URLs by status
-    const statusGroups = source?.domains?.reduce((acc, domain) => {
-      const status =
-        domain.status === "NOT_PROCESSED"
-          ? "NOT_PROCESSED"
-          : domain.status?.toLowerCase() === "fail"
-            ? "FAIL"
-            : "SUCCESS";
-
-      acc[status] = (acc[status] || 0) + 1;
-
-      return acc;
-    }, {});
-
-    const handleBadgeClick = (e, status, source) => {
-      if (isSourcesProcessing) return;
-      e.preventDefault();
-      e.stopPropagation();
-
-      // Set the initial tab first
-      const tabValue =
-        status === "SUCCESS"
-          ? "success"
-          : status === "NOT_PROCESSED"
-            ? "not_processed"
-            : "failed";
-
-      setInitialActiveTab(tabValue);
-
-      // Small delay to ensure state is updated before opening dialog
-      setTimeout(() => {
-        handleEditSource(source, tabValue);
-      }, 0);
-    };
-
-    return (
-      <div className="flex items-center space-x-2">
-        {Object.entries(statusGroups).map(([status, count]) => {
-          if (count === 0) return null;
-
-          let badgeProps = {
-            className:
-              "flex items-center rounded-full gap-1 px-2 py-1 text-body4 font-medium cursor-pointer",
-            variant: "secondary",
-            text: `${count} URL${count > 1 ? "s" : ""}`
-          };
-
-          switch (status) {
-            case "SUCCESS":
-              badgeProps = {
-                ...badgeProps,
-                icon: LinkIcon,
-                iconColor: "text-blue-base",
-                className: `${badgeProps.className} hover:bg-blue-50`
-              };
-              break;
-            case "NOT_PROCESSED":
-              badgeProps = {
-                ...badgeProps,
-                icon: AlertTriangle,
-                iconColor: "text-warning-base",
-                className: `${badgeProps.className} hover:bg-warning-50`
-              };
-              break;
-            case "FAIL":
-              badgeProps = {
-                ...badgeProps,
-                icon: AlertTriangle,
-                iconColor: "text-error-base",
-                className: `${badgeProps.className} hover:bg-error-50`
-              };
-              break;
-            default:
-              return null;
-          }
-
-          return (
-            <div
-              key={status}
-              className={cn(
-                "cursor-pointer",
-                isSourcesProcessing && "pointer-events-none opacity-50"
-              )}
-              role="button"
-              tabIndex={0}
-              onClick={(e) => handleBadgeClick(e, status, source)}>
-              <Badge {...badgeProps} />
-            </div>
-          );
-        })}
-      </div>
-    );
   };
 
   // Add cleanup when closing dialogs
@@ -2168,168 +1935,19 @@ export default function NewGuru({ guruData, isProcessing }) {
                 }
               });
             }}>
-            <div className="max-w-md">
-              <FormField
-                control={form.control}
-                name="guruName"
-                render={({ field }) => (
-                  <FormItem>
-                    <div className="flex items-center space-x-2">
-                      <FormLabel>
-                        Guru Name <span className="text-red-500">*</span>
-                      </FormLabel>
-                      <HeaderTooltip
-                        text={
-                          isEditMode
-                            ? "Guru name cannot be changed"
-                            : "Enter the name of your AI guru"
-                        }
-                      />
-                    </div>
-                    <FormControl>
-                      <Input
-                        placeholder="Enter a guru name"
-                        {...field}
-                        className={
-                          isEditMode ||
-                          isProcessing ||
-                          form.formState.isSubmitting
-                            ? "bg-gray-100 cursor-not-allowed"
-                            : ""
-                        }
-                        disabled={
-                          isEditMode ||
-                          isProcessing ||
-                          form.formState.isSubmitting
-                        }
-                        onBlur={() => form.trigger("guruName")}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="guruLogo"
-                render={({ field: { value, onChange, ...rest } }) => (
-                  <FormItem>
-                    <div className="flex items-center space-x-2">
-                      <FormLabel>
-                        Guru Logo <span className="text-red-500">*</span>
-                      </FormLabel>
-                      <HeaderTooltip text={"Guru logo"} />
-                    </div>
-                    <FormControl>
-                      <div className="flex items-center space-x-4">
-                        <div className="w-16 h-16 bg-gray-900 rounded-lg border-[0.5px] border-gray-85 flex items-center justify-center overflow-hidden">
-                          {iconUrl ? (
-                            <Image
-                              alt="Guru logo"
-                              className="w-full h-full object-cover"
-                              height={64}
-                              src={iconUrl}
-                              width={64}
-                            />
-                          ) : selectedFile ? (
-                            <Image
-                              alt="Selected logo"
-                              className="w-full h-full object-cover"
-                              height={64}
-                              src={URL.createObjectURL(selectedFile)}
-                              width={64}
-                            />
-                          ) : (
-                            <div>
-                              <SolarGalleryAddBold
-                                className="text-gray-500"
-                                height={26.66}
-                                width={26.66}
-                              />
-                            </div>
-                          )}
-                        </div>
-                        <div>
-                          <Input
-                            accept="image/png, image/jpeg"
-                            className="hidden"
-                            id="logo-upload"
-                            type="file"
-                            onChange={(e) => {
-                              const file = e.target.files?.[0];
-
-                              if (file) {
-                                setSelectedFile(file);
-                                setIconUrl(null);
-                                onChange(file);
-                                // Mark form as dirty when logo changes
-                                setDirtyChanges((prev) => ({
-                                  ...prev,
-                                  guruUpdated: true
-                                }));
-                              }
-                            }}
-                            {...rest}
-                            value="" // Always keep the value empty for file inputs
-                          />
-                          <Button
-                            className="text-gray-600 hover:text-gray-700 mb-2"
-                            disabled={
-                              isProcessing || form.formState.isSubmitting
-                            }
-                            type="button"
-                            variant="outline"
-                            onClick={() =>
-                              document.getElementById("logo-upload").click()
-                            }>
-                            <Upload className="mr-2 h-4 w-4" />{" "}
-                            {isEditMode ? "Change Logo" : "Upload Logo"}
-                          </Button>
-                          <FormDescription>
-                            We support PNG, JPEG under 1MB.
-                          </FormDescription>
-                        </div>
-                      </div>
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="guruContext"
-                render={({ field }) => (
-                  <FormItem>
-                    <div className="flex items-center space-x-2">
-                      <FormLabel>
-                        Topics <span className="text-red-500">*</span>
-                      </FormLabel>
-                      <HeaderTooltip
-                        text={
-                          'Add comma-separated topics related to this Guru, e.g., "programming, microservices, containers". This helps the AI understand the Guru\'s expertise and context.'
-                        }
-                      />
-                    </div>
-                    <FormControl>
-                      <Input
-                        placeholder="Enter topics, separated by commas"
-                        {...field}
-                        className="w-full"
-                        disabled={
-                          isSourcesProcessing ||
-                          isProcessing ||
-                          form.formState.isSubmitting
-                        }
-                        onBlur={() => form.trigger("guruContext")}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
+            {/* Use GuruDetailsSection component */}
+            <GuruDetailsSection
+              form={form}
+              isEditMode={isEditMode}
+              isProcessing={isProcessing}
+              isSubmitting={form.formState.isSubmitting}
+              isSourcesProcessing={isSourcesProcessing}
+              iconUrl={iconUrl}
+              selectedFile={selectedFile}
+              setSelectedFile={setSelectedFile}
+              setIconUrl={setIconUrl}
+              setDirtyChanges={setDirtyChanges}
+            />
 
             <div className="max-w-3xl">
               <div className="flex items-center space-x-4">
@@ -2345,249 +1963,27 @@ export default function NewGuru({ guruData, isProcessing }) {
               </div>
             </div>
 
-            {/* Table Area */}
-            <div className="max-w-full">
-              <div className="flex flex-col mb-5">
-                <h3 className="text-lg font-semibold mb-1">Sources</h3>
-                <div className="flex items-center justify-between">
-                  <p className="text-body2 text-gray-400">
-                    Your guru will answer questions based on the sources you
-                    provided below
-                  </p>
-                </div>
-              </div>
+            {/* Use SourcesTableSection component */}
+            <SourcesTableSection
+              sources={sources}
+              isProcessing={isProcessing}
+              isSourcesProcessing={isSourcesProcessing}
+              isSubmitting={form.formState.isSubmitting}
+              isLoadingIntegration={isLoadingIntegration}
+              jiraIntegration={jiraIntegration}
+              fileInputRef={fileInputRef}
+              handleEditSource={handleEditSource}
+              handleDeleteSource={handleDeleteSource}
+              handleReindexSource={handleReindexSource}
+              handlePrivacyBadgeClick={handlePrivacyBadgeClick}
+              setClickedSource={setClickedSource} // Pass needed setters
+              setIsYoutubeSidebarOpen={setIsYoutubeSidebarOpen}
+              setIsJiraSidebarOpen={setIsJiraSidebarOpen}
+              setIsUrlSidebarOpen={setIsUrlSidebarOpen}
+              setShowJiraIntegrationModal={setShowJiraIntegrationModal}
+              // isSourceProcessing={isSourceProcessing} // Pass isSourceProcessing
+            />
 
-              {/* Table Header Actions */}
-              <div className="flex items-center justify-between space-x-4 mb-3">
-                {sources.length > 0 && (
-                  <Select
-                    disabled={
-                      isSourcesProcessing ||
-                      isProcessing ||
-                      form.formState.isSubmitting
-                    }
-                    onValueChange={(value) => setFilterType(value)}>
-                    <SelectTrigger className="guru-sm:w-[100px] guru-md:w-[180px] guru-lg:w-[180px]">
-                      <SelectValue placeholder="All" />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="all">All</SelectItem>
-                      <SelectItem value="website">Website</SelectItem>
-                      <SelectItem value="youtube">Video</SelectItem>
-                      <SelectItem value="pdf">Files</SelectItem>
-                      <SelectItem value="jira">Jira</SelectItem>
-                    </SelectContent>
-                  </Select>
-                )}
-                <SourceActions
-                  isProcessing={isProcessing}
-                  isSourcesProcessing={isSourcesProcessing}
-                  isLoadingIntegration={isLoadingIntegration}
-                  jiraIntegration={jiraIntegration}
-                  onAddYoutubeClick={() => {
-                    setClickedSource([]);
-                    setIsYoutubeSidebarOpen(true);
-                  }}
-                  onAddJiraClick={() => {
-                    setClickedSource([]);
-                    setIsJiraSidebarOpen(true);
-                  }}
-                  onAddWebsiteClick={() => {
-                    setClickedSource([]);
-                    setIsUrlSidebarOpen(true);
-                  }}
-                  onUploadPdfClick={() => {
-                    fileInputRef.current.click();
-                  }}
-                  setShowJiraIntegrationModal={setShowJiraIntegrationModal}
-                />
-              </div>
-
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className="w-[20%]">Type</TableHead>
-                    <TableHead className="w-[50%]">Name</TableHead>
-                    <TableHead className="w-[30%]"></TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {sources.length === 0 ? (
-                    <TableRow>
-                      <TableCell className="text-center" colSpan={3}>
-                        No sources added
-                      </TableCell>
-                    </TableRow>
-                  ) : (
-                    (() => {
-                      const filteredSources =
-                        filterType === "all"
-                          ? sources
-                          : sources.filter(
-                              (source) =>
-                                source.type.toLowerCase() ===
-                                filterType.toLowerCase()
-                            );
-
-                      const urlSources = filteredSources.filter(
-                        (source) =>
-                          source.type.toLowerCase() === "youtube" ||
-                          source.type.toLowerCase() === "website" ||
-                          source.type.toLowerCase() === "jira"
-                      );
-                      const fileSources = filteredSources.filter(
-                        (source) => source.type.toLowerCase() === "pdf"
-                      );
-
-                      const groupedSources = urlSources.reduce(
-                        (acc, source) => {
-                          const domain = getNormalizedDomain(source.url);
-
-                          if (!domain) return acc;
-
-                          const existingSource = acc.find(
-                            (item) => item.domain === domain
-                          );
-
-                          if (existingSource) {
-                            existingSource.count += 1;
-                            existingSource.domains.push(source);
-                          } else {
-                            acc.push({
-                              ...source,
-                              count: 1,
-                              domains: [source],
-                              domain: domain
-                            });
-                          }
-
-                          return acc;
-                        },
-                        []
-                      );
-
-                      const displaySources = [
-                        ...groupedSources,
-                        ...fileSources
-                      ];
-
-                      return displaySources.map((source) => (
-                        <TableRow key={source.id}>
-                          <TableCell className="font-medium">
-                            <div className="flex items-center">
-                              {source.type?.toLowerCase() === "website" && (
-                                <LinkIcon className="mr-2 h-4 w-4" />
-                              )}
-                              {source.type?.toLowerCase() === "youtube" && (
-                                <SolarVideoLibraryBold className="mr-2 h-4 w-4" />
-                              )}
-                              {source.type?.toLowerCase() === "pdf" && (
-                                <SolarFileTextBold className="mr-2 h-4 w-4" />
-                              )}
-                              {source.type?.toLowerCase() === "jira" && (
-                                <JiraIcon className="mr-2 h-4 w-4" />
-                              )}
-                              <span>{source.sources}</span>
-                            </div>
-                          </TableCell>
-
-                          <TableCell>
-                            {isSourceProcessing(source) &&
-                            isSourcesProcessing ? (
-                              <div className="flex items-center gap-2 text-gray-500">
-                                <LoaderCircle className="h-4 w-4 animate-spin" />
-                                <span className="text-sm">
-                                  Processing source...
-                                </span>
-                              </div>
-                            ) : source?.domain?.length > 50 ||
-                              source?.name?.length > 50 ? (
-                              <TooltipProvider>
-                                <Tooltip>
-                                  <TooltipTrigger asChild>
-                                    <span>
-                                      {source?.type?.toLowerCase() === "pdf"
-                                        ? source?.name?.slice(0, 50)
-                                        : source?.domain?.slice(0, 50)}
-                                    </span>
-                                  </TooltipTrigger>
-                                  <TooltipContent>
-                                    <p>
-                                      {source?.type?.toLowerCase() === "pdf"
-                                        ? source?.name
-                                        : source?.domain}
-                                    </p>
-                                  </TooltipContent>
-                                </Tooltip>
-                              </TooltipProvider>
-                            ) : (
-                              <span>
-                                {source?.type?.toLowerCase() === "pdf"
-                                  ? source?.name
-                                  : source?.domain}
-                              </span>
-                            )}
-                          </TableCell>
-
-                          <TableCell className="">
-                            <div className="flex items-center space-x-2 justify-end">
-                              {renderBadges(source)}
-                              <span>
-                                <DropdownMenu>
-                                  <DropdownMenuTrigger asChild>
-                                    <Button
-                                      className="h-8 w-8 p-0"
-                                      disabled={isSourcesProcessing}
-                                      size="icon"
-                                      variant="ghost">
-                                      <span className="sr-only">Open menu</span>
-                                      <MoreVertical className="h-4 w-4" />
-                                    </Button>
-                                  </DropdownMenuTrigger>
-                                  <DropdownMenuContent align="end">
-                                    {(source.type.toLowerCase() === "website" ||
-                                      source.type.toLowerCase() === "youtube" ||
-                                      source.type.toLowerCase() === "jira") && (
-                                      <DropdownMenuItem
-                                        disabled={isSourcesProcessing}
-                                        onClick={() =>
-                                          handleEditSource(source)
-                                        }>
-                                        <Edit className="mr-2 h-3 w-3" />
-                                        Edit
-                                      </DropdownMenuItem>
-                                    )}
-                                    {source.type.toLowerCase() ===
-                                      "website" && (
-                                      <DropdownMenuItem
-                                        disabled={isSourcesProcessing}
-                                        onClick={() =>
-                                          handleReindexSource(source)
-                                        }>
-                                        <RotateCw className="mr-2 h-3 w-3" />
-                                        Reindex
-                                      </DropdownMenuItem>
-                                    )}
-                                    <DropdownMenuItem
-                                      disabled={isSourcesProcessing}
-                                      onClick={() =>
-                                        handleDeleteSource(source)
-                                      }>
-                                      <SolarTrashBinTrashBold className="mr-2 h-3 w-3" />
-                                      Delete
-                                    </DropdownMenuItem>
-                                  </DropdownMenuContent>
-                                </DropdownMenu>
-                              </span>
-                            </div>
-                          </TableCell>
-                        </TableRow>
-                      ));
-                    })()
-                  )}
-                </TableBody>
-              </Table>
-            </div>
             <div className="w-full">
               {hasFormChanged && isEditMode && !isUpdating && (
                 <PendingChangesIndicator
